+++ CHANGELOG.md
@@ -44,7 +44,8 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Fixed
 
-- Nothing.
+- [#89](https://github.com/zendframework/zend-filter/pull/89) fixes infinite
+  loop on malformed HTML comments in StripTags filter.
 
 ## 2.9.1 - 2018-12-17